chore: add restrictions (decorators, room ownership) to pages

+++ base/views.py
@@ -1,7 +1,9 @@
 # from multiprocessing import context
-from urllib.request import Request
+# from urllib.request import Request
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
@@ -9,7 +11,6 @@ from .models import Room, Topic
 from .forms import RoomForm
 
 # Create your views here.
-# from django.http import HttpResponse
 
 # rooms = [
 #   {'id': 1, 'name': 'Lets learn python'},
@@ -19,12 +20,15 @@ from .forms import RoomForm
 
 def loginPage(request):
 
+  if request.user.is_authenticated:
+    return redirect('home')
+
   if request.method == 'POST':
     username = request.POST.get('username')
     password = request.POST.get('password')
 
     try:
-      user = User.objects.get(username=username) #no real need to instatiate user in this line as it will be overwritten after
+      user = User.objects.get(username=username) #no real need to instatiate user in this line as this instance only exists here this once
     except:
       messages.error(request, 'Username or password does not exist')
     
@@ -39,9 +43,11 @@ def loginPage(request):
   context = {}
   return render(request, 'base/login_register.html', context)
 
+
 def logoutUser(request):
   logout(request)
   return redirect('home')
+
 
 def home(request):
   # q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -61,11 +67,14 @@ def home(request):
   context = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
   return render(request, 'base/home.html', context)
 
+
 def room(request, pk):
   room = Room.objects.get(id=pk)
   context = {'room': room}
   return render(request, 'base/room.html', context)
 
+
+@login_required(login_url='login')
 def createRoom(request):
   form = RoomForm()
   if request.method == 'POST':
@@ -79,9 +88,14 @@ def createRoom(request):
   context = {'form': form}
   return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='login')
 def updateRoom(request, pk):
   room = Room.objects.get(id=pk)
   form = RoomForm(instance=room) # prefils the form
+
+  if request.user != room.host:
+    return HttpResponse('You are not the owner!')
 
   if request.method == 'POST':
     form = RoomForm(request.POST, instance = room) #specifying which room to update with the data submitted in the request
@@ -92,8 +106,14 @@ def updateRoom(request, pk):
   context = {'form': form}
   return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='login')
 def deleteRoom(request, pk):
   room = Room.objects.get(id=pk)
+
+  if request.user != room.host:
+    return HttpResponse('You are not the owner!')
+
   if request.method == 'POST':
     room.delete()
     return redirect('home')
